🚚 api: move usecase to results context
