+++ api/src/certification/results/domain/usecases/get-clea-certified-candidate-by-session.js
